Renamed directory to users
