+++ app/src/androidTest/java/comcmput301f16t01/github/carrier/ElasticUserControllerTest.java
@@ -6,8 +6,8 @@ package comcmput301f16t01.github.carrier;
 
 import java.util.concurrent.ExecutionException;
 
-import comcmput301f16t01.github.carrier.User.User;
-        import comcmput301f16t01.github.carrier.User.UserController;
+import comcmput301f16t01.github.carrier.Users.User;
+        import comcmput301f16t01.github.carrier.Users.UserController;
 
 public class ElasticUserControllerTest extends ApplicationTest {
     UserController uc = new UserController();
