+++ app/src/main/java/comcmput301f16t01/github/carrier/Requests/RiderRequestActivity.java
@@ -34,7 +34,7 @@ import java.util.Locale;
 
 import comcmput301f16t01.github.carrier.FareCalculator;
 import comcmput301f16t01.github.carrier.R;
-import comcmput301f16t01.github.carrier.User.UsernameTextView;
+import comcmput301f16t01.github.carrier.Users.UsernameTextView;
 
 /**
  * This will help us show the request from the perspective of a rider
