+++ app/src/main/java/comcmput301f16t01/github/carrier/UserList.java
@@ -2,7 +2,7 @@ package comcmput301f16t01.github.carrier;
 
 import java.util.ArrayList;
 
-import comcmput301f16t01.github.carrier.User.User;
+import comcmput301f16t01.github.carrier.Users.User;
 
 @Deprecated
 public class UserList extends ArrayList {
