+++ app/src/main/java/comcmput301f16t01/github/carrier/Users/ElasticUserController.java
@@ -1,4 +1,4 @@
-package comcmput301f16t01.github.carrier.User;
+package comcmput301f16t01.github.carrier.Users;
 
 import android.os.AsyncTask;
 import android.util.Log;
@@ -9,8 +9,6 @@ import com.searchly.jestdroid.JestDroidClient;
 
 import java.io.IOException;
 
-import comcmput301f16t01.github.carrier.User.User;
-import comcmput301f16t01.github.carrier.User.UserController;
 import io.searchbox.core.DeleteByQuery;
 import io.searchbox.core.DocumentResult;
 import io.searchbox.core.Index;
