+++ app/src/main/java/comcmput301f16t01/github/carrier/Users/UserController.java
@@ -1,4 +1,4 @@
-package comcmput301f16t01.github.carrier.User;
+package comcmput301f16t01.github.carrier.Users;
 
 import android.util.Log;
 import android.util.Patterns;
