+++ app/src/main/res/layout/activity_login.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="comcmput301f16t01.github.carrier.User.LoginActivity"
+    tools:context="comcmput301f16t01.github.carrier.Users.LoginActivity"
     android:background="@color/colorPrimaryDark">
     <!-- TODO add our login background -->
