+++ app/src/main/res/layout/activity_new_vehicle.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:layout_alignParentBottom="true"
-    tools:context="comcmput301f16t01.github.carrier.User.NewUserActivity">
+    tools:context="comcmput301f16t01.github.carrier.Users.NewUserActivity">
 
     <TextView
         android:text="@string/vehiclePrompt"
