+++ app/src/test/java/comcmput301f16t01/github/carrier/RiderTests.java
@@ -3,8 +3,8 @@ package comcmput301f16t01.github.carrier;
 import org.junit.After;
 import org.junit.Test;
 
-import comcmput301f16t01.github.carrier.User.User;
-import comcmput301f16t01.github.carrier.User.UserController;
+import comcmput301f16t01.github.carrier.Users.User;
+import comcmput301f16t01.github.carrier.Users.UserController;
 
 import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertFalse;
