+++ app/src/test/java/comcmput301f16t01/github/carrier/UserProfileTests.java
@@ -3,8 +3,8 @@ package comcmput301f16t01.github.carrier;
 import org.junit.Test;
 
 import comcmput301f16t01.github.carrier.Requests.Request;
-import comcmput301f16t01.github.carrier.User.User;
-import comcmput301f16t01.github.carrier.User.UserController;
+import comcmput301f16t01.github.carrier.Users.User;
+import comcmput301f16t01.github.carrier.Users.UserController;
 
 import static org.junit.Assert.*;
